Move music_engraver.h/cpp to services/
