+++ src/services/music_engraver.cpp
@@ -7,7 +7,7 @@
 
 #include <fullscore/helpers/duration_helper.h>
 #include <fullscore/models/note.h>
-#include <fullscore/music_engraver.h>
+#include <fullscore/services/music_engraver.h>
